// src/app/shared/models/models.ts
export interface SideBarMenu {
  showMyTab?: boolean;
  menuValue?: string;
  menuValue2?:string;
  route?: string;
  hasSubRoute?: boolean;
  showSubRoute: boolean;
  homeName?:string;
  secondData?:boolean;
  icon: string;
  base?: string;
  base2?:string;
  base3?:string;
  base4?:string;
  base5?:string;
  base7?:string;
  base8?:string;
  base9?:string;
  base10?:string;
  last1?: string;
  last?: string;
  page?: string;
  last2?: string;
  count?:string;
  materialicons?: string;
  subMenus: SubMenu[];
  dot?: boolean;
  changeLogVersion?: boolean;
  hasSubRouteTwo?: boolean;
  page1?: string;
  img?: string;
}
export interface breadCrumbItems {
  label: string;
  active?: boolean;
}
export interface SubMenu {
  menuValue: string;
  route?: string;
  base?: string;
  page?: string;
  page1?: string;
  page2?: string;
  base2?: string;
  base3?: string;
  base4?: string;
  base5?: string;
  base6?: string;
  base7?: string;
  base8?: string;
  last?:string;
  dot?:boolean;
  currentActive?: boolean;
  hasSubRoute?: boolean;
  showSubRoute?: boolean;
  customSubmenuTwo?: boolean;
  subMenusTwo?: SubMenu[];
}

export interface SideBar {
  tittle: string;
  subTitle?:string;
  subTitle2?:string;
  base?: string;
  base1?: string;
  base2?: string;
  base3?: string;
  base4?: string;
  base5?: string;
  base6?: string;
  showAsTab: boolean,
  separateRoute: boolean,
  route?: string;
  img?:string;
  twoTitle?:boolean;
  menu: SideBarMenu[];
  navLocation?:string;

}
export interface SideBar2 {
  tittle: string;
  subTitle?:string;
  subTitle2?:string;
  base?: string;
  base1?: string;
  base2?: string;
  base3?: string;
  base4?: string;
  base5?: string;
  base6?: string;
  showAsTab: boolean,
  separateRoute: boolean,
  route?: string;
  img?:string;
  twoTitle?:boolean;
  menu: SideBarMenu[];

}
export interface SubMenuTwo {
  menuValue: string;
  route: string;
  base?: string;
  base1?: string;
  base2?: string;
  base3?: string;
  hasSubRoute?: boolean;
  showSubRoute?: boolean;
}

export interface SubMenu2 {
  menuValue: string;
  route?: string;
  base?: string;
  base1?: string;
  base2?: string;
  base3?: string;
  hasSubRoute?: boolean;
  showSubRoute?: boolean;
  customSubmenuTwo?: boolean;
  subMenusTwo?: SubMenuTwo[];
}
export interface Menu {
  menuValue: string;
  hasSubRouteTwo?: boolean;
  showSubRoute?: boolean;
  hasSubRoute?: boolean;
  icon?: string;
  base?: string;
  base1?: string;
  base2?: string;
  base3?: string;
  base4?: string;
  base5?: string;
  base6?: string;
  base7?: string;
  base8?: string;
  base9?: string;
  base10?: string;
  subMenus?: SubMenu2[];
  customSubmenuTwo?: boolean;
  subMenusTwo?: SubMenuTwo[];
}
export interface MainMenu {
  title: string;
  showAsTab: boolean;
  separateRoute: boolean;
  menu: Menu[];
}